+++ src/main/java/com/ss/ita/kata/implementation/YevgenAleksandrovich/SixImpl.java
@@ -1,6 +1,6 @@
 package com.ss.ita.kata.implementation.YevgenAleksandrovich;
 
-public class Six implements com.ss.ita.kata.Six {
+public class SixImpl implements com.ss.ita.kata.Six {
     @Override
     public long findNb(long m) {
         return 0;
