+++ src/main/java/com/ss/ita/kata/implementation/users/User.java
@@ -8,7 +8,7 @@ public enum User {
     
 	VLADYSLAV_DMYTRIV ("Vladyslav Dmytriv", com.ss.ita.kata.implementation.vladdmytriv.KataHolder.getInstance()),
 	GRANT_VLAD("Grand Vlad", com.ss.ita.kata.implementation.VladGranat.KataHolder.getInstance()),
-	HUZAR_DNITRY("Huzar Dmitry", com.ss.ita.kata.implementation.ErrDmitry404.KataHolder.getInstance()),
+	HUZAR_DMITRY("Huzar Dmitry", com.ss.ita.kata.implementation.ErrDmitry404.KataHolder.getInstance()),
 	VOLODYMYR_KOSTYRA("Volodymyr Kostyra", com.ss.ita.kata.implementation.Percifalll.KataHolder.getInstance()),
 	TARAS_KRASITSKYI("Krasitskyi Taras", com.ss.ita.kata.implementation.TarasKr.KataHolder.getInstance()),
 	YURIY_KURIAN("Kurian Yurii", com.ss.ita.kata.implementation.NyxLex.KataHolder.getInstance()),
